fix: syn.scanner now uses a dedicated pcap handle to prevent deadlocks and improve performances
